Modularize clipboard history
