+++ java/src/org/futo/inputmethod/latin/uix/actions/ClipboardHistoryUtils.kt
@@ -1,0 +1,13 @@
+package org.futo.inputmethod.latin.uix.actions
+
+fun String.toFNV1aHash(): Long {
+    val fnvPrime: Long = 1099511628211L
+    var hash: Long = -3750763034362895579L
+
+    for (byte in this.toByteArray()) {
+        hash = hash xor byte.toLong()
+        hash *= fnvPrime
+    }
+
+    return hash
+}
